Edited Activity 01_h
Combined my solutions.

Now it will compute what are the prime numbers within the range given when the numbers are not in the ArrayList. It will get the Prime Numbers of the range from the arrayList if the range is 2 up to 1223.

According to what I've read 1 is not a prime number...

based on Sir Mac's feedback :)